examples/schemapile/pyproject.toml
[project]
name = "dbxcarta-schemapile-example"
version = "0.1.0"
description = "Reference dbxcarta preset that materializes SchemaPile slices as Delta tables"
authors = [{ name = "Ryan Knight" }]
requires-python = ">=3.12"
dependencies = [
    "dbxcarta>=0.2.32",
    "databricks-sdk>=0.40",
    "python-dotenv",
    "orjson",
]

[project.optional-dependencies]
test = ["pytest"]

[project.scripts]
dbxcarta-schemapile-slice = "dbxcarta_schemapile_example.slice_runner:main"
dbxcarta-schemapile-select = "dbxcarta_schemapile_example.candidate_selector:main"
dbxcarta-schemapile-bootstrap = "dbxcarta_schemapile_example.bootstrap:main"
dbxcarta-schemapile-materialize = "dbxcarta_schemapile_example.materialize:main"
dbxcarta-schemapile-generate-questions = "dbxcarta_schemapile_example.question_generator:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
dbxcarta = { path = "../..", editable = true }

[tool.pytest.ini_options]
addopts = "-m 'not live'"
markers = [
    "live: tests that require a live Databricks profile, a foundation-model endpoint, and a previously-bootstrapped catalog (default pytest skips these)",
]
